[Decomp][Unit] Cohesion split: extract spell/heal/melee damage-bonus + crit to UnitSpellBonus.cpp
